requirements.txt
websockets==12.0
aiohttp==3.9.1
python-telegram-bot==20.7
python-dotenv==1.0.0
loguru==0.7.2
